LayoutTests/traversal/node-iterator-prototype-expected.txt

This test checks whether DOM wrappers created by NodeIterator have their prototypes attached to the correct objects. These nodes are from the child frame, so that's where their prototypes should be attached.

test: child PASS
A: child PASS
B: child PASS
C: child PASS
D: child PASS
D: child PASS
